requirements.txt
bcrypt==4.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==9.3.0
Pygments==2.19.1
rich==14.0.0
